% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Convierte Temperatura de Celsius a Kelvin}
\usage{
celsius_to_kelvin(temperatura_celsius)
}
\arguments{
\item{temperatura_celsius}{Una temperatura en grado Celsius.}
}
\value{
La temperatura ingresada pero en grados Kelvin.
}
\description{
La función celsius_to_Kelvin toma una temperatura en grados Celsius y la convierte en grados Kelvin.
}
